Initial support with some support for edge weighting and edge directional state
