--- GraphPlanarityTesting.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.168
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphPlanarityTesting", "GraphPlanarityTesting\GraphPlanarityTesting.csproj", "{B92C56E0-7A00-49A3-B058-C8A4428344BF}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphPlanarityTesting.Tests", "GraphPlanarityTesting.Tests\GraphPlanarityTesting.Tests.csproj", "{939E90FC-2695-4292-9961-1F4087FAB194}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B92C56E0-7A00-49A3-B058-C8A4428344BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B92C56E0-7A00-49A3-B058-C8A4428344BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B92C56E0-7A00-49A3-B058-C8A4428344BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B92C56E0-7A00-49A3-B058-C8A4428344BF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{939E90FC-2695-4292-9961-1F4087FAB194}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{939E90FC-2695-4292-9961-1F4087FAB194}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{939E90FC-2695-4292-9961-1F4087FAB194}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{939E90FC-2695-4292-9961-1F4087FAB194}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4EAE6EC1-E59C-49C5-B07C-996477CB684F}
-	EndGlobalSection
-EndGlobal
