--- GraphPlanarityTesting/GraphPlanarityTesting.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <Authors>Ondřej Nepožitek</Authors>
-    <Company>Ondřej Nepožitek</Company>
-    <Description>Implementation of the Boyer-Myrvold algorithm for planarity testing of graphs.</Description>
-    <RepositoryUrl>https://github.com/OndrejNepozitek/GraphPlanarityTesting</RepositoryUrl>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-	<PackageLicenseExpression>MIT</PackageLicenseExpression>
-  </PropertyGroup>
-
-</Project>
